Split arg parsing between headless and GUI clients.
