+++ src/main/java/games/strategy/engine/framework/headlessGameServer/HeadlessGameServer.java
@@ -17,6 +17,7 @@ import games.strategy.engine.chat.Chat;
 import games.strategy.engine.chat.IChatPanel;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.properties.GameProperties;
+import games.strategy.engine.framework.ArgParser;
 import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.ServerGame;
 import games.strategy.engine.framework.startup.launcher.ILauncher;
@@ -35,6 +36,7 @@ import games.strategy.triplea.Constants;
 import games.strategy.util.MD5Crypt;
 import games.strategy.util.ThreadUtil;
 import games.strategy.util.TimeManager;
+import games.strategy.util.Util;
 
 /**
  * A way of hosting a game, but headless.
@@ -648,12 +650,125 @@ public class HeadlessGameServer {
   }
 
   public static void main(final String[] args) {
-    GameRunner.handleCommandLineArgs(args, getProperties(), GameRunner.GameMode.HEADLESS_BOT);
+    System.getProperties().setProperty(GameRunner.TRIPLEA_HEADLESS, "true");
+    if(!ArgParser.handleCommandLineArgs(args, getProperties(), GameRunner.GameMode.HEADLESS_BOT)) {
+      usage();
+      return;
+    }
+
+    handleHeadlessGameServerArgs();
     ClipPlayer.setBeSilentInPreferencesWithoutAffectingCurrent(true);
     try {
       new HeadlessGameServer();
     } catch (final Exception e) {
       ClientLogger.logError("Failed to start game server: " + e);
     }
+  }
+
+  private static void usage() {
+    System.out.println("\nUsage and Valid Arguments:\n"
+        + "   " + GameRunner.TRIPLEA_GAME_PROPERTY + "=<FILE_NAME>\n"
+        + "   " + GameRunner.TRIPLEA_GAME_HOST_CONSOLE_PROPERTY + "=<true/false>\n"
+        + "   " + GameRunner.TRIPLEA_SERVER_PROPERTY + "=true\n"
+        + "   " + GameRunner.TRIPLEA_PORT_PROPERTY + "=<PORT>\n"
+        + "   " + GameRunner.TRIPLEA_NAME_PROPERTY + "=<PLAYER_NAME>\n"
+        + "   " + GameRunner.LOBBY_HOST + "=<LOBBY_HOST>\n"
+        + "   " + LobbyServer.TRIPLEA_LOBBY_PORT_PROPERTY + "=<LOBBY_PORT>\n"
+        + "   " + GameRunner.LOBBY_GAME_COMMENTS + "=<LOBBY_GAME_COMMENTS>\n"
+        + "   " + GameRunner.LOBBY_GAME_HOSTED_BY + "=<LOBBY_GAME_HOSTED_BY>\n"
+        + "   " + GameRunner.LOBBY_GAME_SUPPORT_EMAIL + "=<youremail@emailprovider.com>\n"
+        + "   " + GameRunner.LOBBY_GAME_SUPPORT_PASSWORD + "=<password for remote actions, such as remote stop game>\n"
+        + "   " + GameRunner.LOBBY_GAME_RECONNECTION + "=<seconds between refreshing lobby connection [min "
+        + GameRunner.LOBBY_RECONNECTION_REFRESH_SECONDS_MINIMUM + "]>\n"
+        + "   " + GameRunner.TRIPLEA_SERVER_START_GAME_SYNC_WAIT_TIME + "=<seconds to wait for all clients to start the game>\n"
+        + "   " + GameRunner.TRIPLEA_SERVER_OBSERVER_JOIN_WAIT_TIME + "=<seconds to wait for an observer joining the game>\n"
+        + "   " + GameRunner.MAP_FOLDER + "=mapFolder"
+        + "\n"
+        + "   You must start the Name and HostedBy with \"Bot\".\n"
+        + "   Game Comments must have this string in it: \"automated_host\".\n"
+        + "   You must include a support email for your host, so that you can be alerted by lobby admins when your "
+        + "host has an error."
+        + " (For example they may email you when your host is down and needs to be restarted.)\n"
+        + "   Support password is a remote access password that will allow lobby admins to remotely take the "
+        + "following actions: ban player, stop game, shutdown server."
+        + " (Please email this password to one of the lobby moderators, or private message an admin on the "
+        + "TripleaWarClub.org website forum.)\n");
+  }
+
+  private static void handleHeadlessGameServerArgs() {
+    boolean printUsage = false;
+    final String playerName = System.getProperty(GameRunner.TRIPLEA_NAME_PROPERTY, "");
+    final String hostName = System.getProperty(GameRunner.LOBBY_GAME_HOSTED_BY, "");
+    if (playerName.length() < 7 || hostName.length() < 7 || !hostName.equals(playerName)
+        || !playerName.startsWith("Bot") || !hostName.startsWith("Bot")) {
+      System.out.println(
+          "Invalid argument: " + GameRunner.TRIPLEA_NAME_PROPERTY + " and " + GameRunner.LOBBY_GAME_HOSTED_BY
+              + " must start with \"Bot\" and be at least 7 characters long and be the same.");
+      printUsage = true;
+    }
+
+    final String comments = System.getProperty(GameRunner.LOBBY_GAME_COMMENTS, "");
+    if (!comments.contains("automated_host")) {
+      System.out.println(
+          "Invalid argument: " + GameRunner.LOBBY_GAME_COMMENTS + " must contain the string \"automated_host\".");
+      printUsage = true;
+    }
+
+    final String email = System.getProperty(GameRunner.LOBBY_GAME_SUPPORT_EMAIL, "");
+    if (email.length() < 3 || !Util.isMailValid(email)) {
+      System.out.println(
+          "Invalid argument: " + GameRunner.LOBBY_GAME_SUPPORT_EMAIL + " must contain a valid email address.");
+      printUsage = true;
+    }
+
+    final String reconnection =
+        System.getProperty(GameRunner.LOBBY_GAME_RECONNECTION, "" + GameRunner.LOBBY_RECONNECTION_REFRESH_SECONDS_DEFAULT);
+    try {
+      final int reconnect = Integer.parseInt(reconnection);
+      if (reconnect < GameRunner.LOBBY_RECONNECTION_REFRESH_SECONDS_MINIMUM) {
+        System.out.println("Invalid argument: " + GameRunner.LOBBY_GAME_RECONNECTION
+            + " must be an integer equal to or greater than " + GameRunner.LOBBY_RECONNECTION_REFRESH_SECONDS_MINIMUM
+            + " seconds, and should normally be either " + GameRunner.LOBBY_RECONNECTION_REFRESH_SECONDS_DEFAULT + " or "
+            + (2 * GameRunner.LOBBY_RECONNECTION_REFRESH_SECONDS_DEFAULT) + " seconds.");
+        printUsage = true;
+      }
+    } catch (final NumberFormatException e) {
+      System.out.println("Invalid argument: " + GameRunner.LOBBY_GAME_RECONNECTION
+          + " must be an integer equal to or greater than " + GameRunner.LOBBY_RECONNECTION_REFRESH_SECONDS_MINIMUM
+          + " seconds, and should normally be either " + GameRunner.LOBBY_RECONNECTION_REFRESH_SECONDS_DEFAULT + " or "
+          + (2 * GameRunner.LOBBY_RECONNECTION_REFRESH_SECONDS_DEFAULT) + " seconds.");
+      printUsage = true;
+    }
+    // no passwords allowed for bots
+    // take any actions or commit to preferences
+    final String clientWait = System.getProperty(GameRunner.TRIPLEA_SERVER_START_GAME_SYNC_WAIT_TIME, "");
+    final String observerWait = System.getProperty(GameRunner.TRIPLEA_SERVER_OBSERVER_JOIN_WAIT_TIME, "");
+    if (clientWait.length() > 0) {
+      try {
+        final int wait = Integer.parseInt(clientWait);
+        GameRunner.setServerStartGameSyncWaitTime(wait);
+      } catch (final NumberFormatException e) {
+        System.out.println(
+            "Invalid argument: " + GameRunner.TRIPLEA_SERVER_START_GAME_SYNC_WAIT_TIME + " must be an integer.");
+        printUsage = true;
+      }
+    }
+    if (observerWait.length() > 0) {
+      try {
+        final int wait = Integer.parseInt(observerWait);
+        GameRunner.setServerObserverJoinWaitTime(wait);
+      } catch (final NumberFormatException e) {
+        System.out.println(
+            "Invalid argument: " + GameRunner.TRIPLEA_SERVER_START_GAME_SYNC_WAIT_TIME + " must be an integer.");
+        printUsage = true;
+      }
+    }
+
+    if (printUsage) {
+      usage();
+      System.exit(-1);
+    }
+
+
   }
 }
